moved files
`Layout.tsx`, `NavMenu.tsx` and `nav-menu.styl` are moved to `/src/components/layout/`
